% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pidiq_calc2.R
\name{pidiq_calc2}
\alias{pidiq_calc2}
\title{Calculate PIDIQ statistics from a plant-immunity elicitation experiment image.}
\usage{
pidiq_calc2(
  img_file,
  output_dir = "./test_pidiq2",
  gsize_cut = NULL,
  test_plt = FALSE,
  col_lab = NULL,
  row_lab = NULL,
  filter = NULL,
  msg = FALSE
)
}
\arguments{
\item{img_file}{Path to plant/plate spray image.}

\item{output_dir}{Path to output directory for storing processed images, if test_plt == TRUE (default: './test_pidiq2').}

\item{gsize_cut}{Minimum segmentation group size threshold for distinguishing plant from
spurious background pixels which pass filtering, if not supplied, then
a segmentation group size distribution plot will be generated to help you
to interactively select the value (default: NULL).}

\item{test_plt}{Plot preliminary segmentation group images (bw and watershed) (default: FALSE).}

\item{col_lab}{For multi-plant images: Supplied labels corresponding to plate/flat column indices (default: NULL).}

\item{row_lab}{For multi-plant images: Supplied labels corresponding to plate/flat row indices (default: NULL).}

\item{filter}{Name of filtering preset corresponding to the 'name' column of the pidiq_presets.csv file,
which is assumed to be found in the current working directory where the script is called.}

\item{msg}{Print progress-messages to console (default: FALSE).}
}
\value{
A data.frame with PIDIQ statistics calculated from the given plant image file.
}
\description{
Calculate PIDIQ statistics from a plant-immunity elicitation experiment image.
}
\examples{
\dontrun{pidiq_calc2()}

}
